requirements.txt
playwright>=1.40.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
slack_sdk>=3.0.0
python-dateutil>=2.8.0
